Move tutorial.py into utilities.testdata and refactor get_local_testdata
