+++ tests/conftest.py
@@ -2,15 +2,15 @@ import pytest
 import xarray as xr
 from xclim.indicators.land._streamflow import fit, stats
 
-from ravenpy.utilities.testdata import get_test_data
+from ravenpy.utilities.testdata import get_local_testdata
 
 
 @pytest.fixture
 def q_sim_1(tmp_path):
     """A file storing a Raven streamflow simulation over one basin."""
-    return get_test_data(
-        "hydro_simulations", "raven-gr4j-cemaneige-sim_hmets-0_Hydrographs.nc"
-    )[0]
+    return get_local_testdata(
+        "hydro_simulations/raven-gr4j-cemaneige-sim_hmets-0_Hydrographs.nc"
+    )
 
 
 @pytest.fixture
