+++ tests/test_data_assimilation.py
@@ -14,13 +14,13 @@ from ravenpy.models import GR4JCN
 from ravenpy.models.rv import RVC
 from ravenpy.models.state import BasinStateVariables
 from ravenpy.utilities.data_assimilation import assimilate, perturbation
-from ravenpy.utilities.testdata import get_test_data
+from ravenpy.utilities.testdata import get_local_testdata
 
 
 def test_perturbation():
-    ts = get_test_data(
-        "raven-gr4j-cemaneige", "Salmon-River-Near-Prince-George_meteo_daily.nc"
-    )[0]
+    ts = get_local_testdata(
+        "raven-gr4j-cemaneige/Salmon-River-Near-Prince-George_meteo_daily.nc"
+    )
     ds = xr.open_dataset(ts)
 
     tmax = ds.tmax.isel(time=slice(0, 10))
@@ -39,9 +39,9 @@ class TestAssimilationGR4JCN:
     def test_simple(self):
 
         # get timeseries
-        ts = get_test_data(
-            "raven-gr4j-cemaneige", "Salmon-River-Near-Prince-George_meteo_daily.nc"
-        )[0]
+        ts = get_local_testdata(
+            "raven-gr4j-cemaneige/Salmon-River-Near-Prince-George_meteo_daily.nc"
+        )
 
         # set number of members. Using 7 here to make it easier to find and debug.
         n_members = 7
