+++ tests/test_raven_multi_model.py
@@ -2,13 +2,13 @@ import datetime as dt
 import zipfile
 
 from ravenpy.models import RavenMultiModel
-from ravenpy.utilities.testdata import get_test_data
+from ravenpy.utilities.testdata import get_local_testdata
 
 
 class TestRavenMultiModel:
     def test_simple(self):
-        ts = get_test_data(
-            "raven-gr4j-cemaneige", "Salmon-River-Near-Prince-George_meteo_daily.nc"
+        ts = get_local_testdata(
+            "raven-gr4j-cemaneige/Salmon-River-Near-Prince-George_meteo_daily.nc"
         )
         model = RavenMultiModel(models=["gr4jcn", "hmets"])
         gr4jcn = (0.529, -3.396, 407.29, 1.072, 16.9, 0.947)
